app/src/main/java/com/nnaroju/newsapp/presentation/home/HomeViewModel.kt
package com.nnaroju.newsapp.presentation.home

import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.cachedIn
import com.nnaroju.newsapp.domain.usecases.news.NewsUseCases
import dagger.hilt.android.lifecycle.HiltViewModel
import javax.inject.Inject

@HiltViewModel
class HomeViewModel @Inject constructor(
    newsUseCases: NewsUseCases
) : ViewModel() {

    val news = newsUseCases.getNews(
        sources = listOf("abc-news", "bbc-news", "business-insider")
    ).cachedIn(viewModelScope)


}